Added payment page with api integration
This is a working payment page with full api and stripe integration.
Variable data is passed to the payment page through the url fragment
which is then queried from the code. The important queries are
"token" and "transaction" which contain the api authentication token
and transaction uuid respectively. From there a
`GET /core/transaction/<uuid>` is called to retreive transaction
data like items purchase, total cost, and notes. Once the customer
supplies their credit card info through Stripe's Checkout, a
`POST /payment/charge/<uuid>` is sent to the server to charge
the customer.