Preparing for transition to s3
The landing page will soon be stored in an s3 bucket instead
of serving from an ec2 instance. The errors.html file is required
by s3 and is only served in the users sends an invalid http request.
The s3deploy.sh script is used to deploy the code to the s3 bucket.